refactor: console views 묶어서 정리
