+++ src/step1-index.js
@@ -3,9 +3,8 @@
  * 브라우저 환경에서 사용하는 css 파일 등을 불러올 경우 정상적으로 빌드할 수 없습니다.
  */
 
+import { Input, Output } from './view/console/index';
 import LottoGame from './controller/LottoGame';
-import Input from './view/Input';
-import Output from './view/Output';
 import { LottoGenerator, StatisticsGenerator, MessageGenerator } from './controller/generator/index';
 
 const views = {
