+++ src/view/console/index.js
@@ -1,0 +1,2 @@
+export { default as Input } from './Input';
+export { default as Output } from './Output';
